Updating __init__.pys of `utilities`.
 - Moved settings from utilities to data_interface.
 - Added TestRequesting.
 - Moved tests from TestUtilities to TestRequesting.
