+++ utilities/__init__.py
@@ -1,3 +1,13 @@
-from utilities.utils.settings import TRIES, FAIL_DELAY, SUCCESS_DELAY
-from utilities.utils.funcs import save_json_file, load_json_file, reformat_json_file
-from utilities import auto_logging
+"""
+This package helps simplify logging, various data types, and json files.
+"""
+
+from utilities.utils import *
+from utilities.auto_logging import *
+
+from_utils = ['flatten_lists', 'load_json_file', 'save_json_file']
+
+from_auto_logging = ['LogLvl', 'set_log_level', 'auto_log', 'logging']
+
+
+__all__ = from_utils + from_auto_logging
