+++ utilities/utils/settings.py
@@ -1,8 +1,3 @@
-# Web Request Defaults
-TRIES: int = 5
-FAIL_DELAY: int = 60
-SUCCESS_DELAY: int = 1
-
 # File Location Defaults
 ENCODING = 'utf-8'
 DATA_DIR_NAME: str = '17LandsData'
